# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: translator-server API
  version: 1.0.0
  summary: Canonical REST API for the Yara translator-server.
  description: |
    The canonical API surface is mounted under `/api/v1`. Legacy aliases under
    `/api/db/*`, `/api/user/*`, `/api/epubs/*`, `/api/backup/*`,
    `/api/browser-workers`, `/api/proxy/*`, `/api/defaults`,
    `/api/translation-jobs/*` are **deprecated** and will be removed on
    **Wed, 01 Jan 2026** (the `Sunset` header is set on every legacy response).
    New clients must use `/api/v1`.

    When the running server and this document diverge, the running server is the
    source of truth.
  contact:
    name: Yara
    url: https://github.com/mfloresz/yara
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
    description: Canonical v1 surface
  - url: /
    description: Server root (legacy aliases live directly under /api/...)
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /novels:
    get:
      tags: [Novels]
      summary: List the user's novels.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/FieldsParam'
        - $ref: '#/components/parameters/SelectParam'
        - name: q
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CollectionEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Novel' }
    post:
      tags: [Novels]
      summary: Create a novel.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NovelCreate' }
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string }
              description: Canonical URL of the new resource, e.g. `/api/v1/novels/<id>`.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Novel' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /novels/{novelId}/chapters:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    get:
      tags: [Chapters]
      summary: List chapters (summaries by default; full when `?includeContent=true`).
      parameters:
        - name: includeContent
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CollectionEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          oneOf:
                            - $ref: '#/components/schemas/ChapterSummary'
                            - $ref: '#/components/schemas/Chapter'
    post:
      tags: [Chapters]
      summary: Upsert a chapter.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChapterUpsert' }
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Chapter' }
  /novels/{novelId}/chapters/eligible:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    get:
      tags: [Chapters]
      summary: Summaries eligible for the given operation.
      parameters:
        - name: operation
          in: query
          schema: { type: string, enum: [translate, refine], default: translate }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CollectionEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/ChapterSummary' }
  /novels/{novelId}/chapter-summaries:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    get:
      tags: [Chapters]
      summary: Lightweight, paginated chapter summaries.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CollectionEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/ChapterSummary' }
  /novels/{novelId}/chapter-stats:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    get:
      tags: [Chapters]
      summary: Aggregate char counts.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ChapterStats' }
  /novels/{novelId}/chapters/gaps:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    get:
      tags: [Chapters]
      summary: Detect missing chapter numbers in the sequence.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          gaps:
                            type: array
                            items: { type: integer, format: int32 }
  /novels/{novelId}/chapters/{chapterId}:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
      - $ref: '#/components/parameters/ChapterIdParam'
    get:
      tags: [Chapters]
      summary: Get one chapter (full record).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Chapter' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Chapters]
      summary: Delete a chapter.
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '404': { $ref: '#/components/responses/NotFound' }
  /novels/{novelId}/chapters/{chapterId}/status:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
      - $ref: '#/components/parameters/ChapterIdParam'
    patch:
      tags: [Chapters]
      summary: Update a chapter's translation status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, processing, completed, failed, skipped]
                errorMessage:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Chapter' }
  /novels/{novelId}/chapters/bulk-delete:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    post:
      tags: [Chapters]
      summary: Delete many chapters.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted: { type: integer }
                          requested: { type: integer }
  /novels/{novelId}/chapters/clean:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    post:
      tags: [Chapters]
      summary: Apply cleaning rules to a list of chapters.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CleanRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CleanSummary' }
  /novels/{novelId}/chapters/clean-preview:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    post:
      tags: [Chapters]
      summary: Preview a clean operation without persisting.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CleanRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CleanPreview' }
  /novels/{novelId}/chapters/clean-preview-bulk:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    post:
      tags: [Chapters]
      summary: Bulk preview a clean operation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CleanRequest'
                - type: object
                  properties:
                    chapterIds:
                      type: array
                      items: { type: string }
      responses:
        '200':
          description: OK

  # ---------------------------------------------------------------------------
  # Jobs
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth.token
  parameters:
...
    NovelIdParam:
      name: novelId
      in: path
      required: true
      schema: { type: string }
    ChapterIdParam:
      name: chapterId
      in: path
      required: true
      schema: { type: string }
...
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPageParam:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200 }
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    FieldsParam:
      name: fields
      in: query
      description: Comma-separated list of allowed fields. Heavy fields (coverPath, glossary, tags, aiOptions, translationOptions, cleanupRules, sourceDescription, targetDescription, notes) are excluded unless requested.
      schema: { type: string }
    SelectParam:
      name: select
      in: query
      description: Compat alias for `fields`.
      schema: { type: string }
  responses:
...
    BadRequest:
      description: Bad request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    SingleEnvelope:
      type: object
      description: Canonical v1 single-resource envelope.
      required: [data]
      properties:
        data: {}
    CollectionEnvelope:
      type: object
      description: Canonical v1 collection envelope.
      required: [data, meta, links]
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/Meta'
        links:
          $ref: '#/components/schemas/Links'
...
    Novel:
      type: object
      description: |
        The full novel record. The list endpoint with `?fields=` returns a subset
        of these properties. Heavy fields: coverPath, glossary, tags, aiOptions,
        translationOptions, cleanupRules, sourceDescription, targetDescription,
        notes.
      required: [id, sourceTitle, status, sourceLanguage, targetLanguage]
      properties:
        id: { type: string }
        ownerId: { type: string }
        sourceLanguage: { type: string }
        targetLanguage: { type: string }
        sourceTitle: { type: string }
        sourceAuthor: { type: string }
        sourceDescription: { type: string }
        sourceSeries: { type: string }
        sourceNumber: { type: string }
        targetTitle: { type: string }
        targetAuthor: { type: string }
        targetDescription: { type: string }
        targetSeries: { type: string }
        targetNumber: { type: string }
        url: { type: string, format: uri }
        customCommands: { type: string }
        status:
          type: string
          enum: [ongoing, completed, hiatus, dropped]
        coverPath: { type: string, description: Thumbnail path; falls back to the full cover. }
        isPublic: { type: boolean }
        chapterCount: { type: integer }
        translatedCount: { type: integer }
        completedCount: { type: integer }
        originalCharCount: { type: integer }
        translatedCharCount: { type: integer }
        refinedCharCount: { type: integer }
        totalCharCount: { type: integer }
        maxChapterOrder: { type: integer }
        lastCheckedAt: { type: [string, 'null'], format: date-time }
        lastCheckNewChapters: { type: integer }
        lastReadAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string }
        updatedAt: { type: string }
        glossary:
          type: array
          items: { type: object }
        tags:
          type: array
          items: { type: string }
        prompts: { type: object }
        aiOptions: { type: object }
        translationOptions: { type: object }
        cleanupRules:
          type: array
          items: { type: object }
        notes: { type: string }
        canUpdate:
          type: boolean
          description: Whether the URL is in the parser catalog. Only set when explicitly requested via `?fields=canUpdate`.
        requiresBrowser:
          type: boolean
          description: Whether the URL needs the browser-worker extension. Only set when explicitly requested via `?fields=requiresBrowser`.
    NovelCreate:
      type: object
      required: [sourceTitle]
      properties:
        sourceLanguage: { type: string }
        targetLanguage: { type: string }
        sourceTitle: { type: string }
        sourceAuthor: { type: string }
        sourceDescription: { type: string }
        sourceSeries: { type: string }
        sourceNumber: { type: string }
        targetTitle: { type: string }
        targetAuthor: { type: string }
        targetDescription: { type: string }
        targetSeries: { type: string }
        targetNumber: { type: string }
        glossary: { type: array, items: { type: object } }
        prompts: { type: object }
        notes: { type: string }
        aiOptions: { type: object }
        translationOptions: { type: object }
        cleanupRules: { type: array, items: { type: object } }
        url: { type: string, format: uri }
        customCommands: { type: string }
        tags:
          type: array
          items: { type: string }
...